reorganizing, adding isotropic mixtures
