+++ pyVBRc/materials/_stiffness.py
